Update README.md
This project is a personal portfolio website designed to showcase my skills, projects, and experience as a web developer. The portfolio highlights my expertise in front-end and back-end technologies, focusing on responsive web design and user experience. It also serves as a professional online presence, allowing potential employers and collaborators to view my work, resume,